refactor(usage): split monolithic fetcher into modular providers
Extract individual AI provider usage logic (including Antigravity,
Claude, Codex, Gemini, GitHub, GLM, Kiro, and MiniMax) into dedicated
provider files under `usage/providers/` and shared utilities under
`shared/`.
