+++ open-sse/services/usage/shared/number.js
@@ -1,0 +1,11 @@
+/**
+ * Coerce a value to a finite number, with fallback.
+ */
+export function toFiniteNumber(value, fallback = 0) {
+  if (typeof value === "number" && Number.isFinite(value)) return value;
+  if (typeof value === "string" && value.trim()) {
+    const parsed = Number(value);
+    if (Number.isFinite(parsed)) return parsed;
+  }
+  return fallback;
+}
