+++ open-sse/services/usage/shared/time.js
@@ -1,0 +1,32 @@
+/**
+ * Parse reset date/time to ISO string
+ * Handles multiple formats: Unix timestamp (ms), ISO date string, etc.
+ */
+export function parseResetTime(resetValue) {
+  if (!resetValue) return null;
+
+  try {
+    if (resetValue instanceof Date) {
+      return resetValue.toISOString();
+    }
+
+    // Unix timestamps from provider APIs may be seconds or milliseconds.
+    if (typeof resetValue === 'number') {
+      return new Date(resetValue < 1e12 ? resetValue * 1000 : resetValue).toISOString();
+    }
+
+    // If it's a numeric string, treat it like a Unix timestamp too.
+    if (typeof resetValue === 'string') {
+      if (/^\d+$/.test(resetValue)) {
+        const timestamp = Number(resetValue);
+        return new Date(timestamp < 1e12 ? timestamp * 1000 : timestamp).toISOString();
+      }
+      return new Date(resetValue).toISOString();
+    }
+
+    return null;
+  } catch (error) {
+    console.warn(`Failed to parse reset time: ${resetValue}`, error);
+    return null;
+  }
+}
